Allow to use custom adb shell command to start activity

+++ src/main/kotlin/com/github/grishberg/performance/CompareFromApkCommands.kt
@@ -11,6 +11,7 @@ import com.github.grishberg.performance.command.LauncherCommand
 import com.github.grishberg.performance.command.ReadLogcatFromReader
 import com.github.grishberg.performance.command.SleepCommand
 import com.github.grishberg.performance.command.StartActivityCommand
+import com.github.grishberg.performance.command.StartActivityCustomAdbShellCommand
 import com.github.grishberg.performance.command.logcat.LogcatParser
 import com.github.grishberg.performance.command.logcat.StartupTimeLogcatParser
 import com.github.grishberg.performance.launcher.DeviceFacade
@@ -24,35 +25,36 @@ private const val SLEEP_AFTER_KILL_SECONDS = 1
 /**
  * Deletes apk with [appId] if installed.
  * Installs apk placed in [firstApk].
- * Executes [measuresCount] times:
+ * Executes [measurementCount] times:
  *      starts activity [startActivityName]
  *      reads logs
  *      kills app.
  *
  * Installs apk placed in [secondApk].
- * Executes [measuresCount] times:
+ * Executes [measurementCount] times:
  *      starts activity [startActivityName]
  *      reads logs
  *      kills app.
  */
 class CompareFromApkCommands(
-        private val appId: String,
-        private val startActivityName: String,
-        private val firstApk: File,
-        private val secondApk: File,
-        private val logger: RunnerLogger,
-        private val measuresCount: Int,
-        private val aggregatorProvider: AggregatorProvider,
-        private val logcatValuesPattern: String,
-        private val stopWordParameterName: String,
-        private val dryRunStopWordParameterName: String,
-        private val permissions: List<String>
+    private val appId: String,
+    private val startActivityName: String,
+    private val startActivityAdbShellCommand: String,
+    private val shouldDeleteBeforeInstall: Boolean,
+    private val firstApk: File,
+    private val secondApk: File,
+    private val logger: RunnerLogger,
+    private val measurementCount: Int,
+    private val aggregatorProvider: AggregatorProvider,
+    private val logcatValuesPattern: String,
+    private val stopWordParameterName: String,
+    private val dryRunStopWordParameterName: String,
+    private val permissions: List<String>
 ) : Commands {
-    private val dryRunLogcatParser = StartupTimeLogcatParser(logger,
-            EmptyMeasurementAggregator,
-            logcatValuesPattern,
-            stopWordParameterName,
-            dryRunStopWordParameterName)
+
+    private val dryRunLogcatParser = StartupTimeLogcatParser(
+        logger, EmptyMeasurementAggregator, logcatValuesPattern, stopWordParameterName, dryRunStopWordParameterName
+    )
 
     override fun execute(device: DeviceFacade, logcatReader: LogcatReader) {
         val commands = buildCommands(logcatReader)
@@ -70,13 +72,15 @@ class CompareFromApkCommands(
         // install first apk
         commands.addAll(installAndPrepareApk(firstApk, logcatReader))
 
-        val logcatParser1 = StartupTimeLogcatParser(logger,
-                aggregatorProvider.aggregator1,
-                logcatValuesPattern,
-                stopWordParameterName,
-                dryRunStopWordParameterName)
+        val logcatParser1 = StartupTimeLogcatParser(
+            logger,
+            aggregatorProvider.aggregator1,
+            logcatValuesPattern,
+            stopWordParameterName,
+            dryRunStopWordParameterName
+        )
 
-        for (i in 0 until measuresCount) {
+        for (i in 0 until measurementCount) {
             commands.add(firstAppMeasurementLoop(i, logcatParser1, logcatReader))
         }
 
@@ -84,13 +88,15 @@ class CompareFromApkCommands(
         // install second apk
         commands.addAll(installAndPrepareApk(secondApk, logcatReader))
 
-        val logcatParser2 = StartupTimeLogcatParser(logger,
-                aggregatorProvider.aggregator2,
-                logcatValuesPattern,
-                stopWordParameterName,
-                dryRunStopWordParameterName)
+        val logcatParser2 = StartupTimeLogcatParser(
+            logger,
+            aggregatorProvider.aggregator2,
+            logcatValuesPattern,
+            stopWordParameterName,
+            dryRunStopWordParameterName
+        )
 
-        for (i in 0 until measuresCount) {
+        for (i in 0 until measurementCount) {
             commands.add(secondAppMeasurementLoop(i, logcatParser2, logcatReader))
         }
         return commands
@@ -99,12 +105,14 @@ class CompareFromApkCommands(
     private fun installAndPrepareApk(apk: File, logcatReader: LogcatReader): List<LauncherCommand> {
         val commands = mutableListOf<LauncherCommand>()
 
-        commands.add(DeleteApkCommand(appId, logger))
+        if (shouldDeleteBeforeInstall) {
+            commands.add(DeleteApkCommand(appId, logger))
+        }
         commands.add(InstallApkCommand(logger, apk))
         if (permissions.isNotEmpty()) {
             commands.add(GrantPermissionsCommand(appId, permissions))
         }
-        commands.add(StartActivityCommand(appId, startActivityName))
+        commands.add(startApkCommand())
         commands.add(ReadLogcatFromReader(logger, dryRunLogcatParser, logcatReader))
         commands.add(SleepCommand(SLEEP_BEFORE_KILL_SECONDS))
         commands.add(KillAppCommand(logger, appId))
@@ -113,23 +121,41 @@ class CompareFromApkCommands(
         return commands
     }
 
-    private fun firstAppMeasurementLoop(iteration: Int, logcatParser: LogcatParser, logcatReader: LogcatReader): CompositeCommand {
-        return CompositeCommand(iteration, logger, listOf(
-                StartActivityCommand(appId, startActivityName),
+    private fun firstAppMeasurementLoop(
+        iteration: Int,
+        logcatParser: LogcatParser,
+        logcatReader: LogcatReader
+    ): CompositeCommand {
+        return CompositeCommand(
+            iteration, logger, listOf(
+                startApkCommand(),
                 ReadLogcatFromReader(logger, logcatParser, logcatReader),
                 SleepCommand(SLEEP_BEFORE_KILL_SECONDS),
                 KillAppCommand(logger, appId),
                 SleepCommand(SLEEP_AFTER_KILL_SECONDS)
-        ))
+            )
+        )
     }
 
-    private fun secondAppMeasurementLoop(iteration: Int, logcatParser: LogcatParser, logcatReader: LogcatReader): CompositeCommand {
-        return CompositeCommand(iteration, logger, listOf(
-                StartActivityCommand(appId, startActivityName),
+    private fun secondAppMeasurementLoop(
+        iteration: Int,
+        logcatParser: LogcatParser,
+        logcatReader: LogcatReader
+    ): CompositeCommand {
+        return CompositeCommand(
+            iteration, logger, listOf(
+                startApkCommand(),
                 ReadLogcatFromReader(logger, logcatParser, logcatReader),
                 SleepCommand(SLEEP_BEFORE_KILL_SECONDS),
                 KillAppCommand(logger, appId),
                 SleepCommand(SLEEP_AFTER_KILL_SECONDS)
-        ))
+            )
+        )
+    }
+
+    private fun startApkCommand() = if (startActivityAdbShellCommand.isNotEmpty()) {
+        StartActivityCustomAdbShellCommand(startActivityAdbShellCommand)
+    } else {
+        StartActivityCommand(appId, startActivityName)
     }
 }
